Five-Hundred-Thirty-Three Commit: Refactor deleteCustomer() in CustomerServices class to handle error if admin delete customer with reviews

+++ src/main/java/com/bookstore/service/CustomerServices.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import com.bookstore.dao.CustomerDAO;
 import com.bookstore.dao.HashGenerator;
+import com.bookstore.dao.ReviewDAO;
 import com.bookstore.entity.Customer;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
@@ -107,10 +108,33 @@ public class CustomerServices {
 	}
 
 	public void deleteCustomer() throws ServletException, IOException {
+		/*
+		 * Integer customerId = Integer.parseInt(request.getParameter("id"));
+		 * customerDAO.delete(customerId); String message =
+		 * "The customer has been deleted successfully!"; listCustomers(message);
+		 */
 		Integer customerId = Integer.parseInt(request.getParameter("id"));
-		customerDAO.delete(customerId);
-		String message = "The customer has been deleted successfully!";
-		listCustomers(message);
+		Customer customer = customerDAO.get(customerId);
+		
+		if (customer != null) {
+			ReviewDAO reviewDAO = new ReviewDAO();
+			long reviewCount = reviewDAO.countByCustomer(customerId);
+			
+			if (reviewCount == 0) {
+				customerDAO.delete(customerId);			
+				String message = "The customer has been deleted successfully.";
+				listCustomers(message);
+			} else {
+				String message = "Could not delete customer with ID " + customerId
+						+ " because he/she posted reviews for books.";
+				CommonUtility.showMessageBackend(message, request, response);
+			}
+		} else {
+			String message = "Could not find customer with ID " + customerId + ", "
+					+ "or it has been deleted by another admin";
+			CommonUtility.showMessageBackend(message, request, response);
+		}
+		
 	}
 
 	private void updateCustomerFieldsFromForm(Customer customer) {
